Discrete prices quantities

+++ src/main/scala/org/economicsl/auctions/package.scala
@@ -20,6 +20,6 @@ package org.economicsl
 package object auctions {
 
   /** Type used to representing currencies. */
-  type Currency = Double  // todo should this be Long or Double?
+  type Currency = Long
 
 }
